allow overriding for specific versions

+++ test/docker/run-docker.sh
@@ -14,14 +14,26 @@ SIMULTANEOUS=4
 # set up with the name "node_dev-<version>"
 NODE_VERSIONS=`cat ${__dirname}/node_versions.list`
 
+if [ $# -gt 0 ] ; then
+  NODE_VERSIONS=$*
+  echo "Using Node versions: ${NODE_VERSIONS}"
+fi
+
 START_TS=`date +%s`
 
 test_node() {
   local OUT=/tmp/nan-test-${NV}.out
   local NV=$1
+  local ID=node_dev-$NV
+
+  docker inspect "$ID" &> /dev/null
+  if [[ $? -ne 0 ]]; then
+    echo -e "\033[31mCould not find container for [\033[1m$NV\033[22m]\033[39m"
+    return
+  fi
 
   # Run test in a Docker container
-  docker run -v ${__dirname}/../../:/nan-src/:ro node_dev-$NV /bin/bash -c " \
+  docker run -v ${__dirname}/../../:/nan-src/:ro $ID /bin/bash -c " \
     rsync -aAXx --delete /nan-src/ /nan/ --exclude .git; \
     cd /nan/test/; \
     npm install; \
